feat: Implement script to copy HTML files to parent directory
Developed a shell script that utilizes the `cp` command to selectively copy
HTML files from the current working directory to its parent directory. The
script ensures that only files that are newer than the versions in the parent
directory or those that do not exist in the parent directory are copied. This
enhances users' ability to manage HTML files across directories and maintain
up-to-date versions of the files.