Add Star Wars IV episode terminal script
This commit adds the 'star_wars_terminal.sh' script to the repository. The script allows users to play the Star Wars Episode IV movie directly in the terminal using the 'telnet' command. By connecting to the server "towel.blinkenlights.nl," the script streams the ASCII art rendition of the movie.