Add '/action' to the PATH as the last directory
This script modifies the PATH environment variable to include '/action' as the
last directory to search for executable programs. By doing so, it ensures that
'/action' is the last location the shell checks when looking for a program,
allowing for custom executable files or scripts to take precedence when present
in '/action'.